fix anthropic http handler
