+++ litellm/llms/custom_httpx/llm_http_handler.py
@@ -149,6 +149,7 @@ class BaseLLMHTTPHandler:
                     logging_obj=logging_obj,
                     data=data,
                     fake_stream=fake_stream,
+                    litellm_params=litellm_params,
                 )
 
             else:
@@ -182,6 +183,7 @@ class BaseLLMHTTPHandler:
                 logging_obj=logging_obj,
                 timeout=timeout,
                 fake_stream=fake_stream,
+                litellm_params=litellm_params,
             )
             return CustomStreamWrapper(
                 completion_stream=completion_stream,
@@ -226,6 +228,7 @@ class BaseLLMHTTPHandler:
         data: str,
         model: str,
         messages: list,
+        litellm_params: dict,
         logging_obj,
         timeout: Optional[Union[float, httpx.Timeout]],
         fake_stream: bool = False,
@@ -260,11 +263,15 @@ class BaseLLMHTTPHandler:
 
         if fake_stream is True:
             completion_stream = provider_config.get_model_response_iterator(
-                streaming_response=response.json(), sync_stream=True
+                streaming_response=response.json(),
+                sync_stream=True,
+                json_mode=litellm_params.get("json_mode", False),
             )
         else:
             completion_stream = provider_config.get_model_response_iterator(
-                streaming_response=response.iter_lines(), sync_stream=True
+                streaming_response=response.iter_lines(),
+                sync_stream=True,
+                json_mode=litellm_params.get("json_mode", False),
             )
 
         # LOGGING
@@ -288,6 +295,7 @@ class BaseLLMHTTPHandler:
         timeout: Union[float, httpx.Timeout],
         logging_obj: LiteLLMLoggingObj,
         data: dict,
+        litellm_params: dict,
         fake_stream: bool = False,
     ):
         completion_stream, _response_headers = await self.make_async_call(
@@ -300,6 +308,7 @@ class BaseLLMHTTPHandler:
             logging_obj=logging_obj,
             timeout=timeout,
             fake_stream=fake_stream,
+            litellm_params=litellm_params,
         )
         streamwrapper = CustomStreamWrapper(
             completion_stream=completion_stream,
@@ -317,6 +326,7 @@ class BaseLLMHTTPHandler:
         headers: dict,
         data: str,
         messages: list,
+        litellm_params: dict,
         logging_obj: LiteLLMLoggingObj,
         timeout: Optional[Union[float, httpx.Timeout]],
         fake_stream: bool = False,
@@ -352,11 +362,15 @@ class BaseLLMHTTPHandler:
             )
         if fake_stream is True:
             completion_stream = provider_config.get_model_response_iterator(
-                streaming_response=response.json(), sync_stream=False
+                streaming_response=response.json(),
+                sync_stream=False,
+                json_mode=litellm_params.get("json_mode", False),
             )
         else:
             completion_stream = provider_config.get_model_response_iterator(
-                streaming_response=response.aiter_lines(), sync_stream=False
+                streaming_response=response.aiter_lines(),
+                sync_stream=False,
+                json_mode=litellm_params.get("json_mode", False),
             )
         # LOGGING
         logging_obj.post_call(
